Rework tag driven release infrastructure
The previous approach of encrypying sensitive.sbt was prone
to leaking passwords to the build log if SBT were to report
an error on a line of code containing a secret.

The commit now switches to encrypting the PGP passphrase and
Sonatype credentials as environment variables. The private key
is still encrypted on disk as it is too large, but now that
we only need to encrypt a single file we can revert to using
the built in `encrypt-file` command in the Travis CI command
line tool.