Added function converter capability based on initial reflection work from Jason Zaugg.
If you have something from java.util.functions, you should be able to

```
import scala.compat.java8.FunctionConverters._
```

and then use `.asScala` to convert to the corresponding Scala function.

If you have a Scala function, use `.asJava` to convert to the best-matching item from java.util.functions.
If you want a function that isn't the best match, use the corresponding `asJavaInterfaceName` method, e.g.

```
asJavaUnaryOperator((i: Int) => i+1)   // Best match would be IntUnaryOperator

The code is created with a code generator that needs to run in Scala 2.11+,
but SBT runs 2.10 by default.  Thus, build.sbtt was reworked to be
multi-module (the sub-module is the code generator).