Tests, manually written, for function converters.
These were written manually to avoid making the same assumptions in code
generation for the converters and the tests, and thereby missing important
bugs.  As a downside, though, they do not have complete code coverage.  This
should not matter given how the code generator works, but it is a weak point.