Automatic widening of numeric types for primitive Streams
The widening conversions follow the usual pattern. Float becomes
Double, Byte/Char/Short become Int. Boolean and Unit are not widened.

For Arrays, where the actual data is unboxed, widening is implemented
through custom Stepper implementations, so you can get e.g. a native
IntStepper for an Array[Byte]. The StreamConverters hierarchy is
modified to prefer these widened primitive steppers, so that, for
example, the `seqStream` extension method on an Array[Byte] returns
an IntStream.

Other collection types with a boxed representation also get the same
widening conversions (so you can still benefit from the better
performance of unboxed streams) but there are no implicitly available
widening steppers. Only the Stream extension methods fall back to
applying a widening stepper in order to produce a widening primitive
stream.

Widening narrow numeric types for streams is consistent with the Streams
API in the JDK. For example, `CharSequence.chars` returns an `IntStream`
instead of a `Stream[Character]`.