Move code back from scala.runtime.java8 to scala.compat.java8
The standard library contains some specialized JFunction* classes in
scala.runtime.java8, so we cannot use the same package here if we ever
want this to work with Java 9 modules. Therefore everything is moved
back to scala.compat.java8 where it was in 0.7.

We have to drop compatibility with Scala 2.12.0-M5 which contains
these classes as part of scala-library in scala.runtime.java8.